pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Vladislav25 <vladkhleborodov@gmail.com>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2.2"
django-admin = "^2.0.2"
python-dotenv = "^1.0.0"
dj-database-url = "^2.0.0"
psycopg2-binary = "^2.9.6"
gunicorn = "^20.1.0"
django-bootstrap4 = "^23.1"
django-crispy-forms = "^2.0"
crispy-bootstrap4 = "^2022.1"
django-filter = "^23.2"


[tool.poetry.group.dev.dependencies]
ipython = "^8.5.0"
django-extensions = "^3.2.3"
pytest-django = "^4.5.2"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
